split entry point
